% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{densityScores}
\alias{densityScores}
\title{densityScores}
\usage{
densityScores(pcaobj, nscores)
}
\arguments{
\item{pcaobj}{PcaClass Object}

\item{nscores}{Number of Scores}
}
\value{
A vector with the estimate density values
}
\description{
densityScores
}
\examples{
\dontrun{
fdvector<-fdensity( fdaobjMale, pcaobj, bandwith=800)
plot(fdvector, type="p")
}
}
\references{
Delaigle, A. and Hall, P. (2010). Defining probability density for a distribution of random functions. Ann. Statist., 38, 1171-1193.
}

