% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{outliergram}
\alias{outliergram}
\title{OutlierGram - OutlierGram}
\usage{
outliergram(fdaobj, mag = T, factorsh = 1.5, factormg = 1.5, p1 = 1,
  p2 = 0, ...)
}
\arguments{
\item{x:}{matrix with n rows and p columns. Each row represents a curve.}

\item{t:}{vector of length p with the time points at which curves are measured.}

\item{mag:}{whether to look for magnitude outliers (mag=T) or not (mag=F).}

\item{factorsh:}{value of the factor for the detection of shape outlier.}

\item{factormg:}{value of the factor for the detection of magnitude outlier.}

\item{p1:}{at each time point, curves with values in the p1 percentile while be shifted downwards and the (mbd,mei)
will be recalculated in the new position.}

\item{p2:}{at each time point, curves with values in the p2 percentile while be shifted upwards and the (mbd,mei)
will be recalculated in the new position.}

\item{plotting:}{whether to produce a plot with the observed curves and detected outliers and the outliergram
(plotting=T) or not (plotting=F).}

\item{title:}{main title for the graphics with the observed curves and detected outliers.}

\item{ylabel:}{ylabel for the graphics with the observed curves and detected outliers.}

\item{linecol:}{color for line in graphics 1= gray and black, 2= multicolor.}

\item{legend:}{whether to produce a legend in the graphics with the observed curves and detected outliers
to distinguish shape and magnitude outliers.}
}
\value{
The component list is:
\itemize{
 \item{ modified.band.depth}{}
 \item{ modified.epigraph.index}{}
 \item{ dist}{}
 \item{ magnitude.outliers}{}
 \item{ shape.outliers}{}
 \item{ outliers }
}
}
\description{
OutlierGram - OutlierGram
}
\examples{
\dontrun{
## OutlierGram
fdaobjMale<-fdaclass(Gwd$Male,Gwd$Age,c(1,18))
OG<-outliergram(fdaobjMale)
#' OG$shape.outliers
OG$magnitude.outliers
OG$outliers
NewMat<-Mat[,-OG$outliers]
matplot(time, NewMat,type="l",col =color,xlab="",ylab="x(t)",main="Observations",lty=1)
}
}
\references{
{Shape Outlier Detection and Visualisation for Functional Data: the Outliergram. Ana Arribas-Gil and Juan Romo.
Biostatistics, (DOI) 10.1093/biostatistics/kxu006, 2014}
}

